requirements.txt
django>=2.2.0<2.3.0
djangorestframework==3.9.4
